2-strlen.c

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -5,15 +5,16 @@
  * _strlen - return the lengh of a string
  * @s: string to be printed
  * Return: 0
- */
-size_t _strlen(char *s)
+ */ 
+int _strlen(char *s)
 {
-	size_t length = 0;
+	int length = 0;
 	
 	while (*s != '\0')
 	{
 		length++;
 		s++;
 	}
+
 	return (length);
 }
